# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tour Management API",
    "description": "API for Managing Staff, Packages, Bookings, Promotions, and Trash Collections",
    "version": "1.0.0"
  },
  "host": "cse-341-final-kh0d.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/staff": {
      "get": {
        "tags": ["Manage Staff"],
        "summary": "Get all staff",
        "description": "Retrieve a list of all staff members.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Manage Staff"],
        "summary": "Add a new staff member",
        "description": "Create a new staff member in the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userName": { "type": "string", "example": "jdoe" },
                "password": { "type": "string", "example": "password123" },
                "firstName": { "type": "string", "example": "John" },
                "lastName": { "type": "string", "example": "Doe" },
                "email": { "type": "string", "example": "jdoe@example.com" },
                "position": { "type": "string", "example": "Manager" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/packages": {
      "get": {
        "tags": ["Manage Packages"],
        "summary": "Get all packages",
        "description": "Retrieve a list of all packages.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Manage Packages"],
        "summary": "Add a new package",
        "description": "Create a new package in the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "Adventure Package" },
                "continent": { "type": "string", "example": "Asia" },
                "country": { "type": "string", "example": "India" },
                "price": { "type": "number", "example": 1200 },
                "pax": { "type": "number", "example": 4 },
                "inclusion": { "type": "string", "example": "Flights, Hotels, Meals" },
                "description": { "type": "string", "example": "A thrilling adventure package." },
                "availability": { "type": "boolean", "example": true },
                "packageCode": { "type": "string", "example": "ADV123" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/bookings": {
      "get": {
        "tags": ["Manage Bookings"],
        "summary": "Get all bookings",
        "description": "Retrieve a list of all bookings.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Manage Bookings"],
        "summary": "Add a new booking",
        "description": "Create a new booking in the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "packageCode": { "type": "string", "example": "ADV123" },
                "firstName": { "type": "string", "example": "John" },
                "lastName": { "type": "string", "example": "Doe" },
                "email": { "type": "string", "example": "johndoe@example.com" },
                "price": { "type": "number", "example": 1200 },
                "discount": { "type": "number", "example": 100 },
                "paymentStatus": { "type": "string", "example": "Paid" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/promotions": {
      "get": {
        "tags": ["Manage Promotions"],
        "summary": "Get all promotions",
        "description": "Retrieve a list of all promotions.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": ["Manage Promotions"],
        "summary": "Add a new promotion",
        "description": "Create a new promotion in the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "10D/9N IN NORWAY" },
                "continent": { "type": "string", "example": "Europe" },
                "country": { "type": "string", "example": "Norway" },
                "price": { "type": "number", "example": 5000 },
                "pax": { "type": "number", "example": 2 },
                "inclusions": { "type": "string", "example": "all day transfers, accommodation, breakfast, train pass, private tour" },
                "description": { "type": "string", "example": "Royal Palace, Oslo Opera House, Akershus Fortress, Karl Johans Gate" },
                "availability": { "type": "number", "example": 5 },
                "discountRate": { "type": "number", "example": 100 },
                "packageCode": { "type": "string", "example": "EUNRWY-A2025" },
                "startDate": { "type": "string", "example": "04-01-2025" },
                "endDate": { "type": "string", "example": "12-02-2025" }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/staff-trash": {
      "get": {
        "tags": ["Manage Staff Trash"],
        "summary": "Get all staff in trash",
        "description": "Retrieve a list of all staff members in the trash.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/packages-trash": {
      "get": {
        "tags": ["Manage Packages Trash"],
        "summary": "Get all packages in trash",
        "description": "Retrieve a list of all packages in the trash.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/bookings-trash": {
      "get": {
        "tags": ["Manage Bookings Trash"],
        "summary": "Get all bookings in trash",
        "description": "Retrieve a list of all bookings in the trash.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/promotions-trash": {
      "get": {
        "tags": ["Manage Promotions Trash"],
        "summary": "Get all promotions in trash",
        "description": "Retrieve a list of all promotions in the trash.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}